+++ src/main/java/com/gajob/dto/user/LoginDto.java
@@ -1,6 +1,6 @@
-package com.gajob.dto;
+package com.gajob.dto.user;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
+
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 import lombok.AllArgsConstructor;
@@ -14,23 +14,13 @@ import lombok.Setter;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class UserDto {
+public class LoginDto {
 
   @NotNull
   @Size(min = 3, max = 50)
   private String email;
 
-  @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
   @NotNull
   @Size(min = 3, max = 100)
   private String password;
-
-  @NotNull
-  @Size(min = 3, max = 50)
-  private String name;
-
-  @NotNull
-  @Size(min = 3, max = 50)
-  private String nickname;
-
 }
