+++ src/main/java/com/gajob/entity/crawling/News.java
@@ -1,4 +1,4 @@
-package com.gajob.entity;
+package com.gajob.entity.crawling;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
