+++ src/main/java/com/gajob/repository/crawling/NewsRepository.java
@@ -1,6 +1,6 @@
-package com.gajob.repository;
+package com.gajob.repository.crawling;
 
-import com.gajob.entity.News;
+import com.gajob.entity.crawling.News;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface NewsRepository extends JpaRepository<News, Long> {
